SP19-3 Reformat Codebase by integrating RuboCop

+++ app/controllers/teams_controller.rb
@@ -1,5 +1,5 @@
 class TeamsController < ApplicationController
-  before_action :logged_in_user, only: [:index, :new, :create, :destroy]
+  before_action :logged_in_user, only: %i[index new create destroy]
   before_action :valid_viewer, only: [:show]
   before_action :team_leader_or_admin,   only: [:destroy]
   before_action :team_leader, only: [:preference]
@@ -9,65 +9,61 @@ class TeamsController < ApplicationController
     if @team
 
       if @team.preferences_filled?
-        flash[:warning] = "Preferences have already been submitted"
+        flash[:warning] = 'Preferences have already been submitted'
         redirect_to user_path
         return
       end
-      @title = "Preference Selector"
-      @projects = Project.where("approved = ?", true)
+      @title = 'Preference Selector'
+      @projects = Project.where('approved = ?', true)
       render 'preference'
     else
-      flash[:warning] = "You are not yet part of any team"
+      flash[:warning] = 'You are not yet part of any team'
       redirect_to current_user
     end
   end
 
   def index
     if current_user.admin?
-      @counts = Hash.new
-      @leaders = Hash.new
-      @status  = Hash.new
-      @preferences = Hash.new
+      @counts = {}
+      @leaders = {}
+      @status  = {}
+      @preferences = {}
       @teams = Team.all
 
       @teams.each do |team|
-        count = (Relationship.where(team_id:team.id).count)
-        @counts[team.id]  =  count
+        count = Relationship.where(team_id: team.id).count
+        @counts[team.id]  = count
         @leaders[team.id] = team.leader
-        @preferences[team.id] = (team.preferences_filled)? '&#9989;' : "&#10060;"
-        @status[team.id] =  ((Assignment.where(:team_id => team.id).blank?) ? "No" : "Yes")
-        end #end do
+        @preferences[team.id] = team.preferences_filled ? '&#9989;' : '&#10060;'
+        @status[team.id] = (Assignment.where(team_id: team.id).blank? ? 'No' : 'Yes')
+      end # end do
 
-  @sorting = params[:sort]
+      @sorting = params[:sort]
 
-  	if !session.key?(:assignorder)
-		  session[:assignorder] = true
-		end
+      session[:assignorder] = true unless session.key?(:assignorder)
 
-		if @sorting == "assigned"
-			unassigned_teams = @teams.select{|x| @status[x.id] == "No"}
-			assigned_teams = @teams.select{|x| @status[x.id] != "No"}
+      if @sorting == 'assigned'
+        unassigned_teams = @teams.select { |x| @status[x.id] == 'No' }
+        assigned_teams = @teams.reject { |x| @status[x.id] == 'No' }
 
-			if(session[:assignorder] == false)
-							@teams = (unassigned_teams + assigned_teams).paginate(page: params[:page])
-							session[:assignorder] = true
-			else
-							@teams = (assigned_teams + unassigned_teams).paginate(page: params[:page])
-							session[:assignorder] = false
-			end
+        if session[:assignorder] == false
+          @teams = (unassigned_teams + assigned_teams).paginate(page: params[:page])
+          session[:assignorder] = true
+        else
+          @teams = (assigned_teams + unassigned_teams).paginate(page: params[:page])
+          session[:assignorder] = false
+        end
 
-		else
-			@teams = @teams.order(@sorting).paginate(page: params[:page])
-		end
-
-
+      else
+        @teams = @teams.order(@sorting).paginate(page: params[:page])
+      end
 
     else
       @team = current_user.is_member_of
       if @team
         redirect_to @team
       else
-        flash[:warning] = "You are not yet part of any team"
+        flash[:warning] = 'You are not yet part of any team'
         redirect_to current_user
       end
     end
@@ -76,67 +72,61 @@ class TeamsController < ApplicationController
       @team_members = {}
       User.all.each do |user|
         res = Relationship.find_by_user_id(user.id)
-        if res != nil
-          @relationship = Relationship.find_by_user_id(user.id)
-          if @team_members[@relationship.team_id] == nil
-            @team_members[@relationship.team_id] = []
-          end
+        next if res.nil?
 
-          @team_members[@relationship.team_id] << user
+        @relationship = Relationship.find_by_user_id(user.id)
+        if @team_members[@relationship.team_id].nil?
+          @team_members[@relationship.team_id] = []
         end
+
+        @team_members[@relationship.team_id] << user
       end
 
       respond_to do |format|
-        format.xlsx {
+        format.xlsx do
           response.headers[
-              'Content-Disposition'
+            'Content-Disposition'
           ] = "attachment; filename='TeamData.xlsx'"
-        }
-        format.html {render :index}
+        end
+        format.html { render :index }
       end
     end
-
-
   end
 
   def new
-  	@team = Team.new
+    @team = Team.new
   end
 
   def show
-  	@team = Team.find(params[:id])
+    @team = Team.find(params[:id])
     @members = @team.members
 
-	@user_names = Array.new
+    @user_names = []
 
-	User.find_each do |user|
+    User.find_each do |user|
+      if (user.admin == false) && Relationship.find_by_user_id(user.id).nil?
+        full_name = user.lastname + ', ' + user.firstname
+        @user_names << full_name
+    end
 
-        if user.admin==false and Relationship.find_by_user_id(user.id)==nil
-            full_name = user.lastname + ", " + user.firstname
-            @user_names << full_name
-        end
-
-        @user_names = @user_names.sort_by { |word| word.downcase }
-     end
-
+      @user_names = @user_names.sort_by(&:downcase)
+    end
   end
 
-  def data_download
-  end
+  def data_download; end
 
-	def remove
-
-    if current_user.admin? && params[:user_id]==Team.find(params[:team_id]).leader.id.to_s()
-      flash[:danger]  = "Error . . .  You cannot remove the team leader!"
+  def remove
+    if current_user.admin? && params[:user_id] == Team.find(params[:team_id]).leader.id.to_s
+      flash[:danger] = 'Error . . .  You cannot remove the team leader!'
 
     else
       @relationship = Relationship.find_by_user_id(params[:user_id])
       @relationship.destroy
-      flash[:success] = "Remove successful"
+      flash[:success] = 'Remove successful'
     end
 
-		redirect_to teams_path
-	end
+    redirect_to teams_path
+  end
 
   def set_preference
     @team = current_user.is_member_of
@@ -151,49 +141,46 @@ class TeamsController < ApplicationController
     end
   end
 
-	def add_user
+  def add_user
     name_arr = params[:user_name].split(/\s*,\s*/)
-    puts (name_arr)
-		usr = User.find_by(firstname: name_arr[1], lastname: name_arr[0])
-    puts (usr)
-		on_team = Relationship.find_by_user_id(usr.id)
+    puts name_arr
+    usr = User.find_by(firstname: name_arr[1], lastname: name_arr[0])
+    puts usr
+    on_team = Relationship.find_by_user_id(usr.id)
 
-		if on_team != nil
-			flash[:error] = "This user is already on a team"
-			redirect_to teams_path
-    elsif !current_user.admin? && (Relationship.where(team_id:params[:team_id]).count) >= 6
-          flash[:danger]  = "Sorry, this team has already reached the capacity of 6 members. "
-          redirect_to teams_path
+    if !on_team.nil?
+      flash[:error] = 'This user is already on a team'
+      redirect_to teams_path
+    elsif !current_user.admin? && Relationship.where(team_id: params[:team_id]).count >= 6
+      flash[:danger] = 'Sorry, this team has already reached the capacity of 6 members. '
+      redirect_to teams_path
     else
-		relationship = Relationship.new
-		relationship.team_id = params[:team_id].to_s
-		relationship.user_id = usr.id
-		relationship.save
-		flash[:success] = "Successfully added user "+ params[:user_name].to_s+" to team"
-		redirect_to teams_path
+      relationship = Relationship.new
+      relationship.team_id = params[:team_id].to_s
+      relationship.user_id = usr.id
+      relationship.save
+      flash[:success] = 'Successfully added user ' + params[:user_name].to_s + ' to team'
+      redirect_to teams_path
     end
-	end
+  end
 
   def create
     if current_user.teams.count != 0
-      flash[:danger] = "You have already created one team"
+      flash[:danger] = 'You have already created one team'
       redirect_to teams_path
     elsif current_user.is_member_of.present?
-      flash[:danger] = "You are already a member of a team"
+      flash[:danger] = 'You are already a member of a team'
       redirect_to root_url
     else
-  	  @team = Team.new(team_params)
-  	  @team.user_id = current_user.id
-  	  @team.code = ('a'..'z').to_a.shuffle.take(4).join()
+      @team = Team.new(team_params)
+      @team.user_id = current_user.id
+      @team.code = ('a'..'z').to_a.shuffle.take(4).join
 
       if @team.save
 
-        if !current_user.admin?
-          current_user.join_team(@team)
-        end
+        current_user.join_team(@team) unless current_user.admin?
 
-
-        flash[:success] = "Team created successfully"
+        flash[:success] = 'Team created successfully'
         redirect_to @team
       else
         render 'new'
@@ -202,13 +189,13 @@ class TeamsController < ApplicationController
   end
 
   def edit
-  	@team = Team.find(params[:id])
+    @team = Team.find(params[:id])
   end
 
   def update
     @team = Team.find(params[:id])
     if @team.update_attributes(team_params)
-      flash[:success] = "Team updated"
+      flash[:success] = 'Team updated'
       redirect_to @team
     else
       render 'edit'
@@ -218,35 +205,34 @@ class TeamsController < ApplicationController
   def destroy
     Team.find(params[:id]).destroy
 
-    #If this team is deleted, destory its project assginment too!
+    # If this team is deleted, destory its project assginment too!
 
     assignment = Assignment.find_by_team_id(params[:id])
 
-    if !assignment.nil?
-      assignment.destroy
-    end
+    assignment.destroy unless assignment.nil?
 
-    flash[:success] = "Team deleted"
+    flash[:success] = 'Team deleted'
     redirect_to teams_path
   end
 
- private
-    def team_params
-      params.require(:team).permit(:name, :preferences_filled)
-    end
+  private
 
-    def team_leader
-      @team = Team.find(params[:id])
-      redirect_to root_url unless @team.is_leader?(current_user)
-    end
+  def team_params
+    params.require(:team).permit(:name, :preferences_filled)
+  end
 
-    def team_leader_or_admin
-      @team = Team.find(params[:id])
-      redirect_to root_url unless @team.is_leader?(current_user) || current_user.admin?
-    end
+  def team_leader
+    @team = Team.find(params[:id])
+    redirect_to root_url unless @team.is_leader?(current_user)
+  end
 
-    def valid_viewer
-      @team = Team.find(params[:id])
-      redirect_to root_url unless current_user.is_member?(@team) || current_user.admin?
-    end
+  def team_leader_or_admin
+    @team = Team.find(params[:id])
+    redirect_to root_url unless @team.is_leader?(current_user) || current_user.admin?
+  end
+
+  def valid_viewer
+    @team = Team.find(params[:id])
+    redirect_to root_url unless current_user.is_member?(@team) || current_user.admin?
+  end
 end
